+++ sentry-logback/src/test/kotlin/io/sentry/logback/SentryAppenderTest.kt
@@ -3,15 +3,16 @@ package io.sentry.logback
 import ch.qos.logback.classic.Level
 import ch.qos.logback.classic.LoggerContext
 import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.anyOrNull
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
+import io.sentry.ITransportFactory
 import io.sentry.Sentry
 import io.sentry.SentryLevel
 import io.sentry.SentryOptions
 import io.sentry.test.checkEvent
 import io.sentry.transport.ITransport
-import io.sentry.transport.TransportResult
 import java.time.Instant
 import java.time.LocalDateTime
 import java.time.ZoneId
@@ -28,13 +29,15 @@ import org.slf4j.LoggerFactory
 import org.slf4j.MDC
 
 class SentryAppenderTest {
-    private class Fixture(minimumBreadcrumbLevel: Level? = null, minimumEventLevel: Level? = null, val transport: ITransport = mock<ITransport>()) {
+    private class Fixture(minimumBreadcrumbLevel: Level? = null, minimumEventLevel: Level? = null) {
         val logger: Logger = LoggerFactory.getLogger(SentryAppenderTest::class.java)
         val loggerContext = LoggerFactory.getILoggerFactory() as LoggerContext
+        val transportFactory = mock<ITransportFactory>()
+        val transport = mock<ITransport>()
         val utcTimeZone: ZoneId = ZoneId.of("UTC")
 
         init {
-            whenever(transport.send(any())).thenReturn(TransportResult.success())
+            whenever(this.transportFactory.create(any())).thenReturn(transport)
             val appender = SentryAppender()
             val options = SentryOptions()
             options.dsn = "http://key@localhost/proj"
@@ -42,7 +45,7 @@ class SentryAppenderTest {
             appender.setMinimumBreadcrumbLevel(minimumBreadcrumbLevel)
             appender.setMinimumEventLevel(minimumEventLevel)
             appender.context = loggerContext
-            appender.setTransport(transport)
+            appender.setTransportFactory(transportFactory)
             val rootLogger = loggerContext.getLogger(Logger.ROOT_LOGGER_NAME)
             rootLogger.level = Level.TRACE
             rootLogger.addAppender(appender)
@@ -66,19 +69,18 @@ class SentryAppenderTest {
 
     @Test
     fun `does not initialize Sentry if Sentry is already enabled`() {
-        val transport = mock<ITransport>()
+        fixture = Fixture()
         Sentry.init {
             it.dsn = "http://key@localhost/proj"
             it.environment = "manual-environment"
-            it.setTransport(transport)
+            it.setTransportFactory(fixture.transportFactory)
         }
-        fixture = Fixture(transport = transport)
         fixture.logger.error("testing environment field")
 
         await.untilAsserted {
             verify(fixture.transport).send(checkEvent { event ->
                 assertEquals("manual-environment", event.environment)
-            })
+            }, anyOrNull())
         }
     }
 
@@ -93,7 +95,7 @@ class SentryAppenderTest {
                 assertEquals("testing message conversion {}, {}", event.message.message)
                 assertEquals(listOf("1", "2"), event.message.params)
                 assertEquals("io.sentry.logback.SentryAppenderTest", event.logger)
-            })
+            }, anyOrNull())
         }
     }
 
@@ -112,7 +114,7 @@ class SentryAppenderTest {
 
                 assertTrue { eventTime.plusSeconds(1).isAfter(utcTime) }
                 assertTrue { eventTime.minusSeconds(1).isBefore(utcTime) }
-            })
+            }, anyOrNull())
         }
     }
 
@@ -124,7 +126,7 @@ class SentryAppenderTest {
         await.untilAsserted {
             verify(fixture.transport).send(checkEvent { event ->
                 assertEquals(SentryLevel.DEBUG, event.level)
-            })
+            }, anyOrNull())
         }
     }
 
@@ -136,7 +138,7 @@ class SentryAppenderTest {
         await.untilAsserted {
             verify(fixture.transport).send(checkEvent { event ->
                 assertEquals(SentryLevel.DEBUG, event.level)
-            })
+            }, anyOrNull())
         }
     }
 
@@ -148,7 +150,7 @@ class SentryAppenderTest {
         await.untilAsserted {
             verify(fixture.transport).send(checkEvent { event ->
                 assertEquals(SentryLevel.INFO, event.level)
-            })
+            }, anyOrNull())
         }
     }
 
@@ -160,7 +162,7 @@ class SentryAppenderTest {
         await.untilAsserted {
             verify(fixture.transport).send(checkEvent { event ->
                 assertEquals(SentryLevel.WARNING, event.level)
-            })
+            }, anyOrNull())
         }
     }
 
@@ -172,7 +174,7 @@ class SentryAppenderTest {
         await.untilAsserted {
             verify(fixture.transport).send(checkEvent { event ->
                 assertEquals(SentryLevel.ERROR, event.level)
-            })
+            }, anyOrNull())
         }
     }
 
@@ -184,7 +186,7 @@ class SentryAppenderTest {
         await.untilAsserted {
             verify(fixture.transport).send(checkEvent { event ->
                 assertNotNull(event.getExtra("thread_name"))
-            })
+            }, anyOrNull())
         }
     }
 
@@ -197,7 +199,7 @@ class SentryAppenderTest {
         await.untilAsserted {
             verify(fixture.transport).send(checkEvent { event ->
                 assertEquals(mapOf("key" to "value"), event.contexts["MDC"])
-            })
+            }, anyOrNull())
         }
     }
 
@@ -209,7 +211,7 @@ class SentryAppenderTest {
         await.untilAsserted {
             verify(fixture.transport).send(checkEvent { event ->
                 assertFalse(event.contexts.containsKey("MDC"))
-            })
+            }, anyOrNull())
         }
     }
 
@@ -227,7 +229,7 @@ class SentryAppenderTest {
                     "maven:sentry-logback" == pkg.name &&
                         BuildConfig.VERSION_NAME == pkg.version
                 })
-            })
+            }, anyOrNull())
         }
     }
 
@@ -252,7 +254,7 @@ class SentryAppenderTest {
                 assertEquals("this should be a breadcrumb #1", breadcrumb.message)
                 assertEquals("io.sentry.logback.SentryAppenderTest", breadcrumb.category)
                 assertEquals(SentryLevel.DEBUG, breadcrumb.level)
-            })
+            }, anyOrNull())
         }
     }
 
@@ -268,7 +270,7 @@ class SentryAppenderTest {
             verify(fixture.transport).send(checkEvent { event ->
                 assertEquals(1, event.breadcrumbs.size)
                 assertEquals("this should be a breadcrumb", event.breadcrumbs[0].message)
-            })
+            }, anyOrNull())
         }
     }
 
@@ -286,7 +288,7 @@ class SentryAppenderTest {
                 assertEquals(2, event.breadcrumbs.size)
                 assertEquals("this should be a breadcrumb", event.breadcrumbs[0].message)
                 assertEquals("this should not be sent as the event but be a breadcrumb", event.breadcrumbs[1].message)
-            })
+            }, anyOrNull())
         }
     }
 
@@ -298,7 +300,7 @@ class SentryAppenderTest {
         await.untilAsserted {
             verify(fixture.transport).send(checkEvent { event ->
                 assertEquals("release from sentry.properties", event.release)
-            })
+            }, anyOrNull())
         }
     }
 }