+++ sentry/src/test/java/io/sentry/AsyncHttpTransportFactoryTest.kt
@@ -5,11 +5,11 @@ import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.test.assertNotNull
 
-class HttpTransportFactoryTest {
+class AsyncHttpTransportFactoryTest {
 
     @Test
     fun `When HttpTransportFactory doesn't have a valid DSN, it throws InvalidDsnException`() {
-        assertFailsWith<InvalidDsnException> { HttpTransportFactory.create(SentryOptions()) }
+        assertFailsWith<InvalidDsnException> { AsyncHttpTransportFactory().create(SentryOptions()) }
     }
 
     @Test
@@ -17,7 +17,7 @@ class HttpTransportFactoryTest {
         val options = SentryOptions().apply {
             dsn = "ttps://key@sentry.io/proj"
         }
-        assertFailsWith<IllegalArgumentException> { HttpTransportFactory.create(options) }
+        assertFailsWith<IllegalArgumentException> { AsyncHttpTransportFactory().create(options) }
     }
 
     @Test
@@ -25,7 +25,7 @@ class HttpTransportFactoryTest {
         val options = SentryOptions().apply {
             dsn = "https://key@sentry.io/proj"
         }
-        val transport = HttpTransportFactory.create(options)
+        val transport = AsyncHttpTransportFactory().create(options)
         assertNotNull(transport)
     }
 }